backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Redis & Caching
redis==5.0.1
aioredis==2.0.1

# Email
emails==0.6.0
jinja2==3.1.2

# File Handling
python-magic==0.4.27
aiofiles==23.2.1

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Optional: OAuth2
authlib==1.2.1
itsdangerous==2.1.2

# Optional: File Storage
boto3==1.34.0
minio==7.2.0
